Linked Maps to Quiz
Functionality has been completed that when typing in the text from QR code, the backend checks if it is valid and goes to either a challenge or quiz, marked in the database
Dynamic urls are in effect to effectively maximise the process. Quizzes automatically fill in with correct information. Points are added to user after submit is pressed